+++ src/Apha.PACT/Apha.PACT.Application/Dtos/TestFeePlanDto.cs
@@ -12,7 +12,7 @@ namespace Apha.PACT.Application.Dtos
         public string TestCode { get; set; } = null!;
         public decimal? UnitPrice { get; set; }
         public double? NoTests { get; set; }
-        public decimal? TestFee { get; set; }
+        public double? TestFee { get; set; }
         public string? Owner { get; set; }
         public int FpsYear { get; set; }
     }
